oops
If you believe this one is stupid, you are right. But checkouts do not
currently work for plugin with Rebar3, so the only way to test is to push. And
of course, you cannot use Dialyzer with the rebar and provider applications
since they are not dependencies. Nothing I can do here.